requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
requests>=2.28
sentence-transformers>=2.2.2
torch
pymupdf
numpy
transformers
google-genai